+++ FitTracker.Api/Eums/Media.cs
@@ -1,0 +1,7 @@
+namespace FitTracker.Api.Enums;
+
+public enum Media
+{
+    Picture,
+    Video
+}